Rename `parse_to_trace` to `emit_trace`.
